requirements-backend.txt
# Core Backend
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1

# LangChain & AI
langchain==0.3.18
langchain-openai==0.3.6
langchain-chroma==0.2.2
langchain-community==0.3.17
langchain-core==0.3.35
langgraph==0.2.74
langsmith==0.3.8

# OpenAI
openai==1.58.1

# Vector Store
chromadb==0.6.3

# Data & Models
pydantic==2.10.6
pydantic-settings==2.7.1

# Web Scraping & Search
duckduckgo_search==7.4.4
beautifulsoup4==4.13.3
lxml==5.3.1

# Utilities
requests==2.32.3
python-docx==1.1.2
Pillow==11.1.0
